fix: better error handling / resuming

+++ src/helpers/savecontent.ts
@@ -3,6 +3,8 @@ import path from 'path'
 import mkdirp from 'mkdirp'
 import chalk from 'chalk'
 
+import { __jobs } from '../services/queue'
+
 // :: ---
 
 const SAVEDIR = process.env.PWD!
@@ -19,13 +21,16 @@ export async function saveContent(content: string, url: URL) {
     ? pathname
     : path.join(pathname, 'index.html')
 
-  const targetfile = path.join(SAVEDIR, filename)
+  const targetfile = path.join(SAVEDIR, filename.replace(/\.page$/, '.html'))
+  if (fs.existsSync(targetfile)) return
+
   await mkdirp(path.dirname(targetfile))
 
   console.log(
     chalk.greenBright('[save]'),
     'saving content for:',
-    chalk.cyanBright(filename)
+    chalk.cyanBright(filename),
+    chalk.yellowBright(`${__jobs.length} jobs pending.`)
   )
 
   await fs.promises.writeFile(targetfile, content)
